cleaned up pipelines for lear
